requirements.txt
asgiref==3.10.0
Django==5.1.4
dnspython==2.8.0
gunicorn==23.0.0
packaging==25.0
pymongo==3.11.4
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
